@@ -13,12 +13,13 @@
     },
     "emscripten": "2.0.29",
     "engines": {
-        "node": ">=8.12.0"
+        "node": ">=12.7.0"
     },
     "workspaces": [
         "rust/*",
         "packages/*",
         "examples/*",
+        "tools/*",
         "docs",
         "cpp/perspective",
         "python/perspective"
@@ -26,34 +27,25 @@
     "devDependencies": {
         "@babel/cli": "^7.8.4",
         "@babel/core": "^7.8.4",
-        "@babel/plugin-proposal-decorators": "^7.8.3",
-        "@babel/plugin-proposal-optional-chaining": "^7.8.3",
-        "@babel/plugin-transform-for-of": "^7.8.4",
         "@babel/preset-env": "^7.8.4",
-        "@rollup/plugin-babel": "^5.2.3",
-        "@rollup/plugin-node-resolve": "^11.1.1",
-        "@rollup/plugin-typescript": "^8.2.5",
         "@types/ws": "^7.2.2",
         "@typescript-eslint/eslint-plugin": "^4.31.0",
         "@typescript-eslint/parser": "^4.31.0",
         "arraybuffer-loader": "^1.0.2",
         "babel-eslint": "^10.0.3",
         "babel-jest": "^25.1.0",
-        "babel-plugin-lodash": "^3.3.4",
-        "babel-plugin-transform-custom-element-classes": "^0.1.0",
-        "base64-font-loader": "0.0.4",
         "chalk": "^2.4.2",
-        "core-js": "^3.6.4",
-        "cpx": "^1.5.0",
+        "cpy-cli": "^3.1.1",
         "css-loader": "^0.28.7",
         "cssnano": "^4.1.10",
         "cssnano-preset-lite": "^1.0.1",
         "dotenv": "^8.1.0",
+        "esbuild": "^0.13.12",
+        "esbuild-plugin-less": "^1.1.5",
         "eslint": "^7.32.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-markdown": "^2.2.0",
         "eslint-plugin-prettier": "^4.0.0",
-        "file-loader": "^6.2.0",
         "fs-extra": "^8.1.0",
         "html-loader": "^0.5.1",
         "html-loader-jest": "^0.2.1",
@@ -72,37 +64,31 @@
         "mkdirp": "^0.5.1",
         "npm-pack-here": "^1.2.0",
         "npm-run-all": "^4.1.3",
-        "postcss": "^8.2.6",
-        "postcss-loader": "^4.1.0",
+        "postcss": "^7.0.19",
+        "postcss-loader": "^4.3.0",
         "pre-commit": "^1.2.2",
         "prettier": "^2.4.0",
         "puppeteer": "^10.2.0",
         "rimraf": "^2.6.2",
-        "rollup": "^2.38.5",
-        "rollup-plugin-filesize": "^9.1.0",
-        "rollup-plugin-postcss": "^4.0.0",
-        "rollup-plugin-sourcemaps": "^0.6.3",
-        "rollup-plugin-string": "^3.0.0",
         "sinon": "^7.3.1",
         "style-loader": "^0.18.2",
         "term-img": "^4.1.0",
-        "ts-jest": "^25.1.0",
         "ts-loader": "^6.2.0",
-        "typedoc": "^0.21.9",
-        "typedoc-plugin-markdown": "^3.10.4",
-        "typedoc-plugin-no-inherit": "^1.3.0",
+        "typedoc": "^0.22.7",
+        "typedoc-plugin-markdown": "^3.11.6",
+        "typedoc-plugin-no-inherit": "^1.3.1",
         "typescript": "4.3.3",
         "webfontloader": "^1.6.28",
         "webpack": "^5.14.0",
         "webpack-cli": "^4.3.1",
         "webpack-dev-server": "^3.11.2",
-        "webpack-fix-style-only-entries": "^0.4.0",
         "worker-loader": "^3.0.7"
     },
     "resolutions": {
         "graceful-fs": "^4.2.4",
         "jest": "^25.1.0",
-        "write-file-atomic": "2.4.1",
+        "write-file-atomic": "3.0.0",
+        "postcss": "^7.0.19",
         "webpack": "^5.14.0",
         "react": "16.8.6",
         "react-dom": "16.8.6",
