@@ -11,7 +11,7 @@ delete process.env.NODE;
 try {
     execSync(`mkdirp ${cwd}`, {stdio});
     process.env.CLICOLOR_FORCE = 1;
-    execSync(`emcmake cmake ${__dirname} -DCMAKE_BUILD_TYPE=${env}`, {
+    execSync(`emcmake cmake ${__dirname} -Wno-dev -DCMAKE_BUILD_TYPE=${env}`, {
         cwd,
         stdio,
     });
@@ -19,8 +19,8 @@ try {
         cwd,
         stdio,
     });
-    execSync(`cpx esm/**/* ../esm`, {cwd, stdio});
-    execSync(`cpx cjs/**/* ../cjs`, {cwd, stdio});
+    execSync(`cpy esm/**/* ../esm`, {cwd, stdio});
+    execSync(`cpy cjs/**/* ../cjs`, {cwd, stdio});
 } catch (e) {
     console.error(e);
     process.exit(1);
