@@ -25,11 +25,7 @@ module.exports = [
                 },
                 {
                     test: /\.(png|jpe?g|gif)$/i,
-                    use: [
-                        {
-                            loader: "file-loader"
-                        }
-                    ]
+                    type: "asset/resource"
                 }
             ]
         },
@@ -50,32 +46,5 @@ module.exports = [
             libraryTarget: "umd",
             path: path.resolve(__dirname, "static/js")
         }
-        // plugins: [new PerspectivePlugin({})],
-        // module: {
-        //     rules: [
-        // {
-        //     test: /\.js$/,
-        //     enforce: "pre",
-        //     use: ["source-map-loader"]
-        // },
-        // {
-        //     test: /\.less$/,
-        //     use: [{loader: "style-loader"}, {loader: "css-loader"}, {loader: "less-loader"}]
-        // },
-        // {
-        //     test: /\.(png|jpe?g|gif)$/i,
-        //     use: [
-        //         {
-        //             loader: "file-loader"
-        //         }
-        //     ]
-        // }
-        //     ]
-        // },
-        // devServer: {
-        //     contentBase: [path.join(__dirname, "dist"), path.join(__dirname, "../../node_modules/superstore-arrow")]
-        // },
-        // devtool: "source-map",
-        // ignoreWarnings: [/Failed to parse source map/]
     }
 ];
