@@ -53,7 +53,7 @@ for (const name of Object.keys(hashes)) {
 
     // Dist
     try {
-        execute_throw`git -C dist/${name} commit -am"Autogenerated"`;
-        execute`git -C dist/${name} push origin master`;
+        // execute_throw`git -C dist/${name} commit -am"Autogenerated"`;
+        // execute`git -C dist/${name} push origin master`;
     } catch (e) {}
 }
