@@ -7,7 +7,7 @@
  *
  */
 
-const {get_config} = require("@finos/perspective/dist/esm/config");
+const {get_config} = require("@finos/perspective");
 const path = require("path");
 const webpack = require("webpack");
 const cssnano = require("cssnano");
@@ -89,14 +89,8 @@ class PerspectiveWebpackPlugin {
         if (!(this.options.inline || this.options.inlineWasm)) {
             rules.push({
                 test: /\.wasm$/,
-                type: "javascript/auto",
                 include: [this.options.wasmPath, this.options.viewerPath],
-                use: {
-                    loader: require.resolve("file-loader"),
-                    options: {
-                        name: this.options.wasmName,
-                    },
-                },
+                type: "asset/resource",
             });
         } else {
             rules.push({
@@ -137,7 +131,7 @@ class PerspectiveWebpackPlugin {
         rules.push({
             test: /\.ttf$/,
             include: /monaco\-editor/,
-            use: [require.resolve("file-loader")],
+            type: "asset/resource",
         });
 
         const perspective_config = get_config();
@@ -163,15 +157,21 @@ class PerspectiveWebpackPlugin {
 
         const plugin_replace = new webpack.NormalModuleReplacementPlugin(
             /@finos\/perspective$/,
-            "@finos/perspective/dist/esm/perspective.parallel.js"
+            "@finos/perspective/dist/esm/perspective.js"
         );
         plugin_replace.apply(compiler);
 
         const plugin_replace2 = new webpack.NormalModuleReplacementPlugin(
             /@finos\/perspective\-viewer$/,
-            "@finos/perspective-viewer/dist/esm/index.js"
+            "@finos/perspective-viewer/dist/esm/perspective-viewer.js"
         );
         plugin_replace2.apply(compiler);
+
+        const plugin_replace3 = new webpack.NormalModuleReplacementPlugin(
+            /@finos\/perspective\-workspace$/,
+            "@finos/perspective-workspace/dist/esm/perspective-workspace.js"
+        );
+        plugin_replace3.apply(compiler);
 
         moduleOptions.rules = (moduleOptions.rules || []).concat(rules);
     }
